revert order utility and disregarded utility calculation changes
