+++ dfusion_rust_core/Cargo.toml
@@ -10,9 +10,9 @@ byteorder = "1.3.1"
 graph = { git = "https://github.com/graphprotocol/graph-node", tag = "v0.13.2" }
 graph-node-reader = { git = "https://github.com/gnosis/graph-node-reader", tag = "v0.13.2" }
 lazy_static = "1.2.0"
+log = "0.4.6"
 mock-it = "0.2.0"
 serde_derive = "1.0"
 serde = "1.0"
 sha2 = "0.8.0"
 web3 = { git = "https://github.com/graphprotocol/rust-web3", branch = "trace-struct-pre-ethereum-types" }
-log = "0.4.6"
