+++ dfusion_rust_core/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod database;
 pub mod models;
-pub mod util;
 
 use graph::prelude::SubgraphDeploymentId;
